fix: 여행정보 제목, 검색 영역 고정
