+++ src/routes/TravelInfo/TravelInfoContainer.tsx
@@ -1,13 +1,9 @@
-import { useEffect } from "react";
+import { Fragment, useEffect } from "react";
 import { useTravelInfo } from "states/useTravelInfo";
 
-import { Bar, Typography } from "@/components";
 import { TravelInfoDetailModal } from "@/components/TravelInfoDetailModal/TravelInfoDetailModal";
-import { cssAlignVerticalStyle } from "@/styles/align";
-import { COLORS } from "@/styles/color";
 
 import { TravelInfoList } from "./components/TravelInfoList";
-import { TravelInfoSearch } from "./components/TravelInfoSearch";
 
 export const TravelInfoContainer = () => {
   const {
@@ -23,22 +19,13 @@ export const TravelInfoContainer = () => {
   }, [setIsOpenTravelInfoDetailModal]);
 
   return (
-    <div css={cssAlignVerticalStyle({ gap: 16 })}>
-      <div css={cssAlignVerticalStyle({ gap: 16, alignItems: "flex-start" })}>
-        <Typography size="20" weight={700} color={COLORS.GRAY3}>
-          여행 정보
-        </Typography>
-        <Bar />
-      </div>
-      <div css={cssAlignVerticalStyle({ gap: 16, alignItems: "flex-start" })}>
-        <TravelInfoSearch />
-        <TravelInfoList />
-      </div>
+    <Fragment>
+      <TravelInfoList />
       <TravelInfoDetailModal
         isOpen={isOpenTravelInfoDetailModal}
         onClose={() => setIsOpenTravelInfoDetailModal(false)}
         travelInfo={selectedContent}
       />
-    </div>
+    </Fragment>
   );
 };
